% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{append_encounter_data}
\alias{append_encounter_data}
\title{Append encounter data to the master import metadata}
\usage{
append_encounter_data(master_metadata, encounter_data)
}
\arguments{
\item{master_metadata}{A data frame representing the master import metadata.}

\item{encounter_data}{A data frame representing the encounter data to be appended.}
}
\value{
A data frame with the encounter data appended to the master metadata.
}
\description{
Append encounter data to the master import metadata
}
\examples{
\dontrun{
updated_master_metadata <- append_encounter_data(master_metadata, encounter_data)
}
}
