% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_master_import_path}
\alias{get_master_import_path}
\title{Get the path of the master import file}
\usage{
get_master_import_path(colony)
}
\arguments{
\item{colony}{A character string specifying the name of the colony.}
}
\value{
A character string representing the path to the master import file.
}
\description{
This function constructs a path to the master import file for a given colony.
}
\examples{

dontrun{
 get_master_import_folder("ColonyA")
}
}
