% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_startup_paths}
\alias{get_startup_paths}
\title{Get paths to all startup Excel files}
\usage{
get_startup_paths()
}
\value{
A character vector containing paths to all Excel files in the specified subdirectory.
}
\description{
This function retrieves paths to all Excel files in the "Starttime files and stored loggers" subdirectory of the sea track folder.
}
\examples{
\dontrun{
all_xlsx_files <- get_startup_paths()
}
}
