% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_unfinished_session}
\alias{get_unfinished_session}
\title{Find a logger's unfinished session in the master startup data frame
This function finds the unfinished session for a given logger in the master startup data frame.}
\usage{
get_unfinished_session(master_startup, logger_id, logger_download_stop_date)
}
\arguments{
\item{master_startup}{A data frame containing the master startup and shutdown information.}

\item{logger_id}{A character string specifying the logger ID.}

\item{logger_download_stop_date}{A Date object specifying the reported download/stop date of the logger.}
}
\value{
A list containing the index of the unfinished session and the session data frame, or NULL if no unfinished session is found.
}
\description{
Find a logger's unfinished session in the master startup data frame
This function finds the unfinished session for a given logger in the master startup data frame.
}
\examples{
\dontrun{
unfinished_session <- get_unfinished_session(master_startup, "Logger123", as.Date("2023-01-15"))
}
}
