% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{handle_partner_metadata}
\alias{handle_partner_metadata}
\title{Add partner provided metadata to the master import file}
\usage{
handle_partner_metadata(
  colony,
  new_metadata,
  master_import,
  nonresponsive_list = list()
)
}
\arguments{
\item{colony}{A character string specifying the name of the colony.}

\item{new_metadata}{List of tibbles, each corresponding to a sheet in the partner provided information file.}

\item{master_import}{List of tibbles, each corresponding to a sheet in the master import file.}
}
\value{
An updated version of the master import file, as a list where each element is a sheet from the excel file.
}
\description{
This function adds metadata provided by partners to a master import file of the appropriate colony.
It firstly adds missing sessions by checking start up files.
It then appends the reported encounter data, avoiding duplicate rows.
Finally it updates sessions based on reported logger returns. This includes generating new sessions for loggers restarted in the field.
}
