% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{handle_returned_loggers}
\alias{handle_returned_loggers}
\title{Handle restarted loggers}
\usage{
handle_returned_loggers(colony, master_startup, logger_returns, restart_times)
}
\arguments{
\item{colony}{A character string specifying the name of the colony.}

\item{master_startup}{A data frame containing the master startup and shutdown information.}

\item{logger_returns}{A data frame containing logger return information.}

\item{restart_times}{A data frame containing logger restart information.}
}
\value{
An updated dataframe containing the modified master import data frame.
}
\description{
This function processes logger return information and updates the master import data frame accordingly.
}
\examples{
dontrun{
updated_master_startup <- handle_returned_loggers(master_startup, logger_returns, restart_times)
}
}
