% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_master_import}
\alias{load_master_import}
\title{Load master import file for a given colony}
\usage{
load_master_import(colony)
}
\arguments{
\item{colony}{A character string specifying the name of the colony.}
}
\value{
A list consisting of two items.
data: A list of tibbles, each corresponding to a sheet in the master import file.
path: The file path of the loaded master import file.
}
\description{
This function loads the master import file for a specified colony.
It iterates through the appropriate sheets and combines the data into a list of data frames.
}
\examples{
\dontrun{
load_master_import("ColonyA")
}
}
