% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_nonresponsive}
\alias{load_nonresponsive}
\title{Load multiple nonresponsive logger sheets}
\usage{
load_nonresponsive(file_paths, manufacturers)
}
\arguments{
\item{file_paths}{A character vector of file paths to load.}

\item{manufacturers}{A character vector of manufacturers, same length as file_paths.}
}
\value{
A named list of tibbles, each containing nonresponsive logger data for the corresponding manufacturer.
}
\description{
This function loads nonresponsive logger sheets for multiple file paths and manufacturers.
}
\examples{
\dontrun{
file_paths <- c("lotek.xlsx", "migratetech.xlsx")
manufacturers <- c("Lotek", "MigrateTech")
nonresponsive_list <- load_nonresponsive(file_paths, manufacturers)
}
}
