% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_partner_metadata}
\alias{load_partner_metadata}
\title{Load partner provided metadata from an Excel file}
\usage{
load_partner_metadata(file_path)
}
\arguments{
\item{file_path}{A character string specifying the path to the Excel file.}
}
\value{
A list of data frames, each corresponding to a sheet in the Excel file.
}
\description{
This function reads metadata provided by partners from an Excel file.
It iterates through the appropriate sheets and combines the data into a list of data frames.
}
\examples{
\dontrun{
load_partner_metadata("path/to/partner_metadata.xlsx")
}
}
