% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_unresponsive}
\alias{load_unresponsive}
\title{Load unresponsive logger sheet for current year
This function loads the record of unresponsive loggers.}
\usage{
load_unresponsive()
}
\value{
A list consisting of two items.
data: The loaded unresponsive logger information.
path: The file path of the loaded unresponsive logger information file.
}
\description{
Load unresponsive logger sheet for current year
This function loads the record of unresponsive loggers.
}
