% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_master_sheet}
\alias{save_master_sheet}
\title{Save a master sheet to an Excel file.}
\usage{
save_master_sheet(new_master_sheets, filepath)
}
\arguments{
\item{new_master_sheets}{A data frame containing the master sheet data to be saved.}

\item{filepath}{A string specifying the path and name of the Excel file to be created.}
}
\value{
No return value.
}
\description{
This function writes the provided data frame (\code{new_master_sheets}) to an Excel file
specified by \code{filename}.
}
\examples{
\dontrun{
save_master_sheet(new_master_sheets, "output.xlsx")
}

}
