% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_nonresponsive}
\alias{save_nonresponsive}
\title{Save multiple nonresponsive logger sheets to Excel files}
\usage{
save_nonresponsive(file_paths, nonresponsive_list)
}
\arguments{
\item{file_paths}{A character vector of file paths to save each sheet.}

\item{nonresponsive_list}{A named list of tibbles, each containing nonresponsive logger data.}
}
\value{
No return value.
}
\description{
This function iterates through a list of nonresponsive logger sheets and a vector of file paths,
saving each sheet to its corresponding file path.
}
\examples{
\dontrun{
save_multiple_nonresponsive(nonresponsive_list, file_paths)
}
}
