% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatrackfunctions-package.R
\docType{package}
\name{seatrackfunctions-package}
\alias{seatrackfunctions}
\alias{seatrackfunctions-package}
\title{seatrackfunctions: Utility Functions for Seatrack}
\description{
Some utility functions to assist in the handling of data as part of the seatrack project
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jevansbio/seatrack_functions}
}

}
\author{
\strong{Maintainer}: Julian Evans \email{julian.evans@nina.no}

}
\keyword{internal}
