% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatrackRmetadata-package.R
\docType{package}
\name{seatrackRmetadata-package}
\alias{seatrackRmetadata}
\alias{seatrackRmetadata-package}
\title{seatrackRmetadata: Utility Functions for Seatrack}
\description{
Some utility functions to assist in the handling of data as part of the seatrack project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ninanor.github.io/seatrack-metadata/}
}

}
\author{
\strong{Maintainer}: Julian Evans \email{julian.evans@nina.no}

}
\keyword{internal}
