% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_comments}
\alias{set_comments}
\title{Set or append comments in the master_startup data frame}
\usage{
set_comments(master_startup, index, logger_comments)
}
\arguments{
\item{master_startup}{A data frame containing a 'comment' column to be updated.}

\item{index}{Integer index specifying the row in master_startup to update.}

\item{logger_comments}{A character string containing the comment to add or append.}
}
\value{
The updated master_startup data frame with the modified comment.
}
\description{
This function updates the 'comment' field of the specified row in the master_startup data frame.
If a non-empty logger comment is provided, it will be set as the comment if no existing comment is present.
If an existing comment is present, the logger comment will be appended to it, separated by " | ".
}
\examples{
master_startup <- data.frame(comment = c("", "Existing comment"))
set_comments(master_startup, 1, "New logger comment")
set_comments(master_startup, 2, "Another logger comment")

}
