% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_master_startup_value}
\alias{set_master_startup_value}
\title{Set a value in a specific cell of master startup}
\usage{
set_master_startup_value(master_startup, index, column, value)
}
\arguments{
\item{master_startup}{Master starup tibble.}

\item{index}{Integer. The row index of the cell to update.}

\item{column}{Character or integer. The column name or index of the cell to update.}

\item{value}{The new value to assign to the specified cell.}
}
\value{
The updated \code{master_startup} data frame.
}
\description{
This function updates the value of a specified cell in the \code{master_startup} data frame.
}
\examples{
set_master_startup_value(master_startup, 2, "download_type", "Succesfully downloaded")

}
