% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_sea_track_folder}
\alias{set_sea_track_folder}
\title{Set the base directory for the sea track folder}
\usage{
set_sea_track_folder(dir)
}
\arguments{
\item{dir}{A character string specifying the path to the base directory.}
}
\value{
None
}
\description{
This function sets a global variable used by other functions.
}
\examples{
dontrun{
 set_sea_track_folder("/path/to/sea/track/folder")
}
}
