% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{start_logging}
\alias{start_logging}
\title{Start logging to a file}
\usage{
start_logging(
  log_dir = NULL,
  log_file = paste0("seatrack_functions_log_", Sys.Date(), ".txt")
)
}
\arguments{
\item{log_dir}{A character string specifying the directory where the log file will be saved. If NULL, the log file will be saved in the current working directory.}

\item{log_file}{A character string specifying the name of the log file. Default is "seatrack_functions_log.txt".}
}
\value{
None
}
\description{
This function initializes logging to a specified directory.
}
\examples{
\dontrun{
start_logging("/path/to/log/directory")
}
}
