% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{handle_returned_loggers}
\alias{handle_returned_loggers}
\title{Handle restarted loggers}
\usage{
handle_returned_loggers(
  colony,
  master_startup,
  logger_returns,
  restart_times,
  nonresponsive_list = list()
)
}
\arguments{
\item{colony}{A character string specifying the name of the colony.}

\item{master_startup}{A data frame containing the master startup and shutdown information.}

\item{logger_returns}{A data frame containing logger return information.}

\item{restart_times}{A data frame containing logger restart information.}

\item{nonresponsive_list}{A list containing tibbles of unresponsive loggers for different manufacturers.
The name of the list element should match the producer name in master_startup (e.g., "Lotek", "MigrateTech").}
}
\value{
A list consisting of two elements:
\itemize{
\item `master_startup``: An updated dataframe containing the modified master import data frame.
\item \code{nonresponsive_list}: An updated list containing the modified nonresponsive logger data frames.
}
}
\description{
This function processes logger return information and updates the master import data frame accordingly.
}
\examples{
dontrun{
updated_master_startup <- handle_returned_loggers(master_startup, logger_returns, restart_times)
}
}
